Account management updated

+++ Account.py
@@ -1,12 +1,19 @@
 import csv
 import os
 import json
+from pathlib import Path
+
+
+mm_dir_path = Path(__file__).parent
 
 class Account:
-    def __init__(self,name):
-        self.balances_lookup = {'Compte courant': 0, 'Girokonto': 1, 'Compte epargne': 2,
-                                'PayPal': 3, 'Geldbeutel': 4, 'Visa': 5}
+    def __init__(self,name,balance,rowsDeleted,colsDeleted,headers,detectString):
         self.name  = name
+        self.balance = balance
+        self.rowsDeleted = rowsDeleted
+        self.colsDeleted = colsDeleted
+        self.headers = headers
+        self.detectString = detectString
         #self.balance = self.getBalance(name)[0]
         #self.date = self.getBalance(name)[1]
 
@@ -15,58 +22,67 @@ class Account:
         self.balance -= amount
         account.balance += amount
 
-    def update(self):
-        self.date = datetime.now()
-        index = self.balances_lookup[self.name]
-        balances[index][0] = round(self.balance,2)
-        balances[index][1] = self.date
-        saveBalances()
 
     def getDate(self):
         return self.date
 
 
-CC_LUX = Account('Compte courant')
-GK_DE = Account('Girokonto')
-PP = Account('PayPal')
-GB = Account('Geldbeutel')
-VISA = Account('Visa')
-CE_LUX = Account('Compte epargne')
+
+
 
 def accountsLookup(account_name):
-    accounts_lookup = {'Compte courant': CC_LUX, 'Girokonto': GK_DE, 'PayPal': PP, 'Geldbeutel': GB, 'Visa': VISA, 'Compte epargne': CE_LUX}
-    try:
-        account = accounts_lookup[account_name]
-    except KeyError:
-        account = None
-    return account
+    for acc in accounts:
+        if acc.name == account_name:
+            account = acc
+            return account
+    return None
 
 
 def statementDetection(file):
     transacts = []
     with open(file,mode="r",encoding="latin-1") as csv_file:
-        csv_reader = csv.reader(csv_file, delimiter=";")
+        dialect = csv.Sniffer().sniff(csv_file.read(1024), delimiters=";,")
+        csv_file.seek(0)
+        csv_reader = csv.reader(csv_file, dialect)
         for index, row in enumerate(csv_reader):
             if len(row) > 0:
                 if row[0][0] >= '0' and row[0][0] <= '9':
                     transacts.append(row)
-    print(transacts)
     return transacts
 
 
 def importAllAccounts(file):
+    accounts = []
     if os.path.isfile(file):
         if os.stat(file).st_size > 0:
             with open(file, "r", encoding="latin-1") as json_file:
-                data = json.load(json_file)
-        else:
-            data = {}
-    else:
-        data = {}
+                account_data = json.load(json_file)
+
+            if "accounts" in account_data:
+                for acc in account_data["accounts"]:
+                    accounts.append(Account(acc["name"], acc["balance"],acc["rowsDeleted"], acc["colsDeleted"], acc["headers"],acc["detectString"]))
+
+    return accounts
+
+def exportAllAccounts(accounts,file):
+    if os.path.isfile(file) and len(accounts)>0:
+        data = {"accounts":[]}
+        for acc in accounts:
+            data['accounts'].append({'name': acc.name,
+                                     'balance': acc.balance,
+                                     'rowsDeleted': acc.rowsDeleted,
+                                     'colsDeleted': acc.colsDeleted,
+                                     'headers': acc.headers,
+                                     'detectString': acc.detectString})
+        with open(file, "w+", encoding="latin-1") as json_file:
+            json_string = json.dumps(data)
+            json_file.write(json_string)
     return data
 
-def deleteAccount(index,data,file):
-    data["accounts"].pop(index)
-    with open(file, "w+", encoding="latin-1") as json_file:
-        json_string = json.dumps(data)
-        json_file.write(json_string)
+def deleteAccount(index,file):
+    accounts.pop(index)
+    exportAllAccounts(accounts,file)
+
+
+accounts_file = mm_dir_path / "accounts.json"
+accounts = importAllAccounts(accounts_file)
